Update dependencies, add Riverpod for state management, and remove unused packages
This commit updates the dependencies to their latest versions, adds Riverpod for state management, and removes unused packages. The `flutter_riverpod` package is upgraded to version 1.0.3, and `provider` is removed as it is no longer needed with Riverpod. The `flutter_localizations` package is added to support localization. The `intl` package is downgraded to version 0.18.1, and `provider` is upgraded to version 6.1.2. Additionally, the `flutter_launcher_icons` configuration is updated to generate launcher icons.